Time: 204 ms (70.97%), Space: 51.2 MB (20.24%) - LeetHub

+++ 128-longest-consecutive-sequence/128-longest-consecutive-sequence.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int longestConsecutive(vector<int>& nums) {
-        map<long long int ,int> m;
+        map<int ,int> m;
         int count=1,ans=1;
         if(nums.size()==0)
             return 0;
